% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{continuousWorker}
\alias{continuousWorker}
\title{The worker function to calculate continuous NKDE (with ARMADILLO and sparse matrix)}
\arguments{
\item{kernel_func}{a cpp pointer function (selected with the kernel name)}

\item{samples_k}{a numeric vector of the actual kernel values, updates at
each recursion}

\item{neighbour_list}{a List, giving for each node an IntegerVector with
its neighbours}

\item{edge_mat}{matrix, to find the id of each edge given two neighbours.}

\item{v}{the actual node to consider for the recursion (int)}

\item{v1}{the connected node to considere for the recursion (int)}

\item{l1}{the edge connecting v and v1 (int)}

\item{d}{the actual distance traveled before the recursion}

\item{alpha}{th actual alpha value before the recursion}

\item{bw}{the kernel bandiwdth}

\item{line_weights}{a vector with the length of the edges}

\item{samples_edgeid}{a vector associating each sample to an edge}

\item{samples_x}{a vector with the x coordinates of each sample}

\item{samples_ya}{vector with the y coordinates of each sample}

\item{nodes_x}{a vector with the x coordinates of each node}

\item{nodes_y}{a vector with the y coordinates of each node}

\item{depth}{the actual recursion depth}

\item{max_depth}{the maximum recursion depth}
}
\value{
a vector with the kernel values calculated for each samples from
the first node given

a vector with the kernel values calculated for each samples from
the first node given
}
\description{
The worker function to calculate continuous NKDE (with ARMADILLO and sparse matrix)

The worker function to calculate continuous NKDE (with ARMADILLO and Integer matrix)
}
