% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{discontinuous_nkde_cpp}
\alias{discontinuous_nkde_cpp}
\title{The main function to calculate discontinuous NKDE}
\usage{
discontinuous_nkde_cpp(
  edge_list,
  neighbour_list,
  events,
  weights,
  samples,
  bw,
  kernel_func,
  nodes,
  line_list,
  max_depth,
  verbose
)
}
\arguments{
\item{edge_list}{A list of edges, accessible by their names}

\item{neighbour_list}{a list of the neighbours of each node}

\item{events}{a numeric vector of the node id of each event}

\item{weights}{a numeric vector of the weight of each event}

\item{samples}{a DataFrame of the samples (with spatial coordinates and belonging edge)}

\item{bw}{the kernel bandwidth}

\item{kernel_func}{the function to calculate the kernel from distance}

\item{nodes}{a DataFrame representing the nodes of the graph (with spatial coordinates)}

\item{line_list}{a DataFrame representing the lines of the graph}

\item{max_depth}{the maximum recursion depth (after which recursion is stopped)}

\item{verbose}{a boolean indicating if the function must print its progress}
}
\value{
a DataFrame with two columns : the kernel values (sum_k) and the number of event reaching each samples (n)
}
\description{
The main function to calculate discontinuous NKDE
}
